chore: begin moving redux logic to shared
