+++ cmd/ui/src/store.ts
@@ -1,17 +1,17 @@
 // Copyright 2023 Specter Ops, Inc.
-// 
+//
 // Licensed under the Apache License, Version 2.0
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-// 
+//
 // SPDX-License-Identifier: Apache-2.0
 
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
@@ -21,6 +21,7 @@ import throttle from 'lodash/throttle';
 import { useDispatch } from 'react-redux';
 import createSagaMiddleware from 'redux-saga';
 import * as reducers from 'src/ducks';
+import { edgeinfo } from 'bh-shared-ui';
 import rootSaga from 'src/rootSaga';
 
 enableMapSet();
@@ -29,6 +30,7 @@ const sagaMiddleware = createSagaMiddleware();
 
 const appReducer = combineReducers({
     ...reducers,
+    edgeinfo,
 });
 
 export const rootReducer = (state: any, action: any) => {
