+++ cmd/ui/src/views/Explore/EdgeInfo/EdgeInfoCollapsibleSection.tsx
@@ -1,25 +1,25 @@
 // Copyright 2023 Specter Ops, Inc.
-// 
+//
 // Licensed under the Apache License, Version 2.0
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-// 
+//
 // SPDX-License-Identifier: Apache-2.0
 
 import { faMinus, faPlus } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Accordion, AccordionDetails, AccordionSummary } from '@mui/material';
+import { EdgeInfoState, EdgeSections, edgeSectionToggle } from 'bh-shared-ui';
 import React, { PropsWithChildren } from 'react';
 import { useSelector } from 'react-redux';
-import { EdgeInfoState, EdgeSections, edgeSectionToggle } from 'src/ducks/edgeinfo/edgeSlice';
 import { AppState, useAppDispatch } from 'src/store';
 import useCollapsibleSectionStyles from 'src/views/Explore/InfoStyles/CollapsibleSection';
 import { SubHeader } from 'src/views/Explore/fragments';
