+++ packages/javascript/bh-shared-ui/package.json
@@ -23,6 +23,7 @@
         "@fortawesome/fontawesome-svg-core": "^1.2.36",
         "@fortawesome/free-solid-svg-icons": "^5.15.4",
         "@fortawesome/react-fontawesome": "^0.1.19",
+        "@reduxjs/toolkit": "^1.9.5",
         "clsx": "^1.1.1",
         "dompurify": "^2.3.0",
         "luxon": "^3.3.0",
