+++ packages/javascript/bh-shared-ui/rollup.config.js
@@ -1,17 +1,17 @@
 // Copyright 2023 Specter Ops, Inc.
-// 
+//
 // Licensed under the Apache License, Version 2.0
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-// 
+//
 // SPDX-License-Identifier: Apache-2.0
 
 import typescript from '@rollup/plugin-typescript';
@@ -37,6 +37,7 @@ export default {
         '@emotion/styled',
         '@fortawesome/free-solid-svg-icons',
         '@fortawesome/react-fontawesome',
+        '@reduxjs/toolkit',
         '@mui/material',
         '@mui/styles',
         '@mui/styles/makeStyles',
