+++ packages/javascript/bh-shared-ui/src/store/edgeinfo/edgeSlice.ts
