+++ packages/javascript/bh-shared-ui/src/store/index.ts
@@ -14,10 +14,5 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-export { default as global } from './global';
-export { default as explore } from './explore';
-export { default as search } from './searchbar';
-export { default as entityinfo } from './entityinfo';
-export { default as tierzero } from './tierzero';
-export { default as assetgroups } from './assetgroups';
-export { default as auth } from './auth/authSlice';
+export * from './edgeinfo/edgeSlice';
+export { default as edgeinfo } from './edgeinfo/edgeSlice';
